// frontend/src/models/emotionalReaction.ts
export interface EmotionalReaction {
    title?: string
    emotions?: string
    what_did_you_do?: string
    what_did_you_think?: string
    when_does_tb_usually_occur?: string
    where_does_tb_occur?: string
    who_is_present_when_tb_occurs?: string
    which_activitie_precede_tb?: string
    wd_other_people_sod_before_tb?: string
    do_you_engage_other_behavior_before_tb_occurs?: string
    what_happens_after_tb?: string
    wdyd_when_tb_occurs?: string
    wd_other_people_do_when_tb_occurs?: string
    what_changes_after_tb_occurs?: string
    wd_you_get_after_tb?: string
}